[Refactoring ⚙️] 커스텀 훅으로 메뉴 클릭 관리
